Moved findCypressWidget
